requirements.txt
discord.py
munch
python-dotenv